+++ changelog.md
@@ -9,6 +9,7 @@
 - Added automatic build with Github Workflows
 - Added posts JSON list for external implementation! Closes #1
 - Edited `projects/htmlAntiCopy`
+- Added "Why?" and "How it works?" section in `projects/htmlAntiCopy`
 
 ## 2023/08/02